scarf::analysis

Struct Control

pub struct Control<'e: 'b, 'b, 'c, A: Analyzer<'e> + 'b> { /* private fields */ }

Provides access to analysis state when in a Analyzer callback.

The most common use cases are:

Access to the current ExecutionState through exec_state, as well as convenience shortcuts:

resolve

resolve_mem

resolve_apply_constraints

read_memory

move_resolved

move_unresolved

Ability to control what branches are analyzed with:

end_branch

end_analysis

add_branch_with_current_state

clear_unchecked_branches

clear_all_branches

Easy way to create analysis for child function with:

analyze_with_current_state

inline

Implementations

impl<'e: 'b, 'b, 'c, A: Analyzer<'e> + 'b> Control<'e, 'b, 'c, A>

pub fn ctx(&self) -> OperandCtx<'e>

Retreives the OperandCtx that was used to construct FuncAnalysis.

pub fn binary( &self, ) -> &'e BinaryFile<<A::Exec as ExecutionState<'e>>::VirtualAddress>

Retreives the BinaryFile that was used to construct FuncAnalysis.

pub fn address(&self) -> <A::Exec as ExecutionState<'e>>::VirtualAddress

For operation() callback, eeturns address of the current instruction that is used as a source for the Operation.

For branch_start() callback, returns the start address of the branch.

For branch_end() callback, returns the address of final instruction of branch.

pub fn branch_start(&self) -> <A::Exec as ExecutionState<'e>>::VirtualAddress

Returns start address of the current branch.

pub fn current_instruction_end( &self, ) -> <A::Exec as ExecutionState<'e>>::VirtualAddress

Returns address of the next instruction that will be executed.

Does not return a valid value in branch_start() callback.

For branch_end() callback, returns the address of instruction after final instruction of the branch.

pub fn end_analysis(&mut self)

Causes the FuncAnalysis to end analysis immediately after this hook is returned from.

Currently no-op in branch_end() callback.

pub fn end_branch(&mut self)

Causes the FuncAnalysis to end the currently simulated branch without executing any remaining instructions or adding branches from Operation::Jump.

Using end_branch together with add_branch_with_current_state for each jump destination when seeing Operation::Jump can be used to customize state which will be used for branches after the jump. It is a bit awkward way to do it and could use a better designed API though :)

Currently no-op in branch_end() callback.

pub fn skip_operation(&mut self)

Skips the current operation.

When used with branches (hopefully?) behaves as if branch is never taken, use end_branch() to also stop analyzing the fallthrough case.

pub fn exec_state(&mut self) -> &mut A::Exec

Returns mutable refence to the current branch’s ExecutionState.

Control includes convenience functions for the most commonly used ExecutionState functions, which can be used instead of calling this.

pub fn user_state(&mut self) -> &mut A::State

Returns mutable reference to user-defined analyzer state Analyzer::State.

pub fn resolve(&mut self, val: Operand<'e>) -> Operand<'e>

Convenience function for exec_state().resolve()

pub fn resolve_mem(&mut self, val: &MemAccess<'e>) -> MemAccess<'e>

Convenience function for exec_state().resolve_mem()

pub fn resolve_register(&mut self, register: u8) -> Operand<'e>

Convenience function for exec_state().resolve_register()

pub fn resolve_flag(&mut self, flag: Flag) -> Operand<'e>

Convenience function for exec_state().resolve_flag()

pub fn resolve_apply_constraints(&mut self, val: Operand<'e>) -> Operand<'e>

Convenience function for exec_state().resolve_apply_constraints()

pub fn read_memory(&mut self, mem: &MemAccess<'e>) -> Operand<'e>

Convenience function for exec_state().read_memory().

pub fn write_memory(&mut self, mem: &MemAccess<'e>, value: Operand<'e>)

Convenience function for exec_state().write_memory().

Sets memory at resolved address mem to value.

pub fn unresolve(&mut self, val: Operand<'e>) -> Option<Operand<'e>>

Don’t use this, will be deprecated at some point. See ExecutionState::unresolve.

pub fn move_unresolved(&mut self, dest: &DestOperand<'e>, value: Operand<'e>)

Convenience function for exec_state().move_to(), with a better name that reminds that this function uses unresolved operands.

pub fn move_resolved(&mut self, dest: &DestOperand<'e>, value: Operand<'e>)

Convenience function for exec_state().move_resolved(),

pub fn set_register(&mut self, register: u8, value: Operand<'e>)

Convenience function for exec_state().set_register().

Equivalent to ctrl.move_resolved(&DestOperand::Register64(register), value).

pub fn set_flag(&mut self, flag: Flag, value: Operand<'e>)

Convenience function for exec_state().set_flag().

Equivalent to ctrl.move_resolved(&DestOperand::Register64(register), value).

pub fn analyze_with_current_state<A2: Analyzer<'e, State = A::State, Exec = A::Exec>>( &mut self, analyzer: &mut A2, entry: <A::Exec as ExecutionState<'e>>::VirtualAddress, )

Takes current analysis’ state as starting state for a function. (“Calls the function”) However, this does not update the state to what this function changes it to.

pub fn inline<A2: Analyzer<'e, State = A::State, Exec = A::Exec>>( &mut self, analyzer: &mut A2, entry: <A::Exec as ExecutionState<'e>>::VirtualAddress, )

Calls the function and updates own state (Both ExecutionState and custom) to a merge of states at this child function’s return points.

NOTE: User is expected to call ctrl.skip_operation() if this is during hook for Operation::Call; this is not done automatically as inlining during other operations is also allowed, even if less useful. (Maybe this is a bit of poor API?)

If the child analyzer uses end_analysis or end_branch to prevent the inlined analyzer reaching Operation::Return, this will end up merging only branches that do return.

If any of the branches reaches a return, the state will be updated even if end_analysis is used later.

pub fn cast<'d, B>(&'d mut self) -> Control<'e, 'b, 'd, B>
where B: Analyzer<'e, Exec = A::Exec, State = A::State>,

Casts to Control<B: Analyzer> with compatible states.

This is fine as the control doesn’t use Analyzer for anything, the type is just defined to take Analyzer as Control is cleaner than Control<Self::Exec, Self::State> when implementing trait methods.

pub fn add_branch_with_current_state( &mut self, address: <A::Exec as ExecutionState<'e>>::VirtualAddress, )

Adds a branch to be analyzed using current state.

pub fn add_branch_with_state( &mut self, address: <A::Exec as ExecutionState<'e>>::VirtualAddress, exec_state: A::Exec, user_state: A::State, )

Adds a branch to be analyzed using given ExecutionState and user-defined state.

pub fn continue_at_address( &mut self, address: <A::Exec as ExecutionState<'e>>::VirtualAddress, )

Causes the current branch jump to an address.

Internally ends the current branch, and queues a new one at the given address using current state.

Slightly differs from what ctrl.end_branch() does in that if used on middle of non-jumping instruction, all Operations in the instruction will be executed before branch ends and execution will be continued at address.

Can be used on jump to only force one branch to be taken. As with branches in general, if the branch has been already analyzed and the new state doesn’t differ from old, the branch doesn’t get analyzed.

pub fn clear_all_branches(&mut self)

Clears all branches, both analyzed and currently pending.

Similar to clear_unchecked_branches, but also clears all branches that had been seen. Effectively same as if a new FuncAnalysis had been started from start of current branch, with the state that the branch was reached at.

pub fn clear_unchecked_branches(&mut self)

Clears any pending (unchecked) branches.

Effectively guarantees that any branches that were seen from Operation::Jump but have not been checked, will not be checked, and only code reachable from the currently executing branch will be checked.

pub fn const_word_offset(&self, left: Operand<'e>, right: u32) -> Operand<'e>

Convenience for cases where address + CONST * REG_SIZE is needed

pub fn mem_word(&self, addr: Operand<'e>, offset: u64) -> Operand<'e>

Either ctx.mem32() or ctx.mem64(), depending on ExecutionState word size.

pub fn mem_access_word(&self, addr: Operand<'e>, offset: u64) -> MemAccess<'e>

ctx.mem_access(addr, offset, E::WORD_SIZE)

pub fn if_mem_word<'a>(&self, op: Operand<'e>) -> Option<&'e MemAccess<'e>>

Either op.if_mem32() or op.if_mem64(), depending on ExecutionState word size.

pub fn if_mem_word_offset<'a>( &self, op: Operand<'e>, offset: u64, ) -> Option<Operand<'e>>