scarf::analysis

Struct FuncAnalysis

pub struct FuncAnalysis<'a, Exec: ExecutionState<'a>, State: AnalysisState> { /* private fields */ }

The analysis runner of scarf.

FuncAnalysis executes through all paths (‘branches’) of a function, calling user-defined callbacks in Analyzer for every state-modifying operation, while keeping track of what branches have been already executed.

State merging and loops

When two separate branches of execution converge, FuncAnalysis will merge the ExecutionState of branches: Any Operand in registers or in memory that differs will be replaced with OperandType::Undefined if it wasn’t Undefined already. If this caused anything in ExecutionState to change, the following branch will be analyzed again, even if analysis had already walked through its instruction once.

Scarf does not have any special understanding of loops, a jump backwards is just another merge point. This effectively often makes scarf to provide user code an ExecutionState that reports registers to contain constants and jumps to be always/never taken for first iteration of a loop, and later on run the branch ‘correctly’ with Undefined when the jump backwards has been seen.

Example

The following code is a strlen implementation in x86-64 assembly. Register rcx has the null-terminated input string, and rax will contain the string length on return.

entry:
    mov rax, 0
loop:
    movzx rdx, byte [rcx]
    add rax, 1
    add rcx, 1
    cmp rdx, 0
    jne loop
exit:
    sub rax, 1 ; Account for null byte having been counted above
    ret

Scarf considers there to 3 branches:

Branch from entry to exit

Branch from loop to exit

Branch from exit that returns

Note that both branch 1 and 2 include the loop instructions. In this case, the backwards jump can only be known to be possibly taken after the first branch has already been simulated, so there would have been no way for scarf to know that movzx rdx, byte [rcx] will be a start of another branch. For consistency, even if a jump to current instruction has been seen, scarf will walk each branch until it sees a jump or return instruction.

On analysis start, scarf will execute branch 1, seeing that on jump the execution splits to two new branches. Resolving rax during this will return constant 0, later on 1, and resolving the jump condition at the end of the branch gives (Mem8[rcx] == 0) == 0 - “Jump if first byte of argument rcx is nonzero”.

After the first branch, both branches 2 and 3 are queued; there is no guarantee which will be analyzed first. If branch 3 is first, scarf will report the function to return (resolving rax) constant 0. Once branch 2 is analyzed, it will have similar ‘known’ results what branch 1 shows (rax changes from 1 to 2, jump condition is (Mem8[rcx + 1] == 0) == 0), as this is the first time branch has started from there.

As branch 2 analysis reaches the jump, scarf queues branches 2 and 3, and notices that the values in registers rax, rcx, and rdx differ from what the older branches 2 and 3 had. Each of those registers will be assigned an Undefined value for the new queued branches. If branch 3 wasn’t analyzed yet, the state with Undefined overwrites the earlier queued ‘will-return-zero’ state.

Finally, scarf walks through branches 2 and 3 with the undefined state (in either order). At end of this second run of branch 2, rax will have value Undefined_a + 1, rcx has Undefined_b + 1, and rdx has Mem8[Undefined_b]. These do not add anything over the states where these registers were Undefined, so no more branches will be queued.

Since branch 3 contains sub rax, 1 instruction, the final return value reported by scarf is the not-very-helpful Undefined - 1.

Other merge details

In general, Operand merging is defined as “If old value is undefined or differs from new, merge to unique undefined value”. However, there are few details that are worth noting:

Undefined values allocated during FuncAnalysis::analyze call will get reused after the call returns. This is mainly worth noting if user code returns results as Operand, and compares them to values from other analysis runs.

If old or new value is undefined, the merged value may recycle either of those values, or allocate a new unique undefined.

When values in memory are merged, a single unit of Operand to be merged is not a byte, but a word (4 or 8 bytes, depending on the ExecutionState). This means that if code has byte-sized globals or structure fields, they may be assigned Undefined because values neighbour address differed. So far this has rarely ended up causing issues, but it may sometimes need special handling from user code.

If the address operand of memory contains Undefined at all, the merge operation may drop the value entirely, reverting it back to ‘original’.

That is, while a write of 5 to Mem32[Undefined + 4] is guaranteed to resolve to 5 in same branch as the write, later branches may resolve just to Mem32[Undefined + 4] again.

Implementations

impl<'a, Exec: ExecutionState<'a>> FuncAnalysis<'a, Exec, DefaultState>

pub fn new( binary: &'a BinaryFile<Exec::VirtualAddress>, ctx: OperandCtx<'a>, start_address: Exec::VirtualAddress, ) -> FuncAnalysis<'a, Exec, DefaultState>

Creates a new FuncAnalysis with default ExecutionState and no custom user-side state.

pub fn with_state( binary: &'a BinaryFile<Exec::VirtualAddress>, ctx: OperandCtx<'a>, start_address: Exec::VirtualAddress, state: Exec, ) -> FuncAnalysis<'a, Exec, DefaultState>

Creates a new FuncAnalysis with user-given ExecutionState and no custom user-side state.

impl<'a, Exec: ExecutionState<'a>, State: AnalysisState> FuncAnalysis<'a, Exec, State>

pub fn custom_state( binary: &'a BinaryFile<Exec::VirtualAddress>, ctx: OperandCtx<'a>, start_address: Exec::VirtualAddress, exec_state: Exec, analysis_state: State, ) -> FuncAnalysis<'a, Exec, State>

Creates a new FuncAnalysis with user-given ExecutionState and custom user-side state.

pub fn analyze<A: Analyzer<'a, State = State, Exec = Exec>>( &mut self, analyzer: &mut A, )

Runs the analysis, calling the user-defined callbacks defined on Analyzer trait as the code is being stepped thorugh.

pub fn finish(self) -> Cfg<'a, Exec, State>

Runs analyzer to end without any user interaction, producing Cfg

Calling this is not necessary if you do not need a Cfg.

Currently if Control::end_analysis has been used to stop the analysis, any remaining branches (but not the one that was being analyzed on end_analysis call!) will be walked through to gain a ‘better’ idea of the control flow graph. This is somewhat inconsistent behaviour and should be changed to not analyze any pending branches once all users relying on this have migrated.

pub fn finish_with_changes<F>(self, hook: F) -> Cfg<'a, Exec, State>
where F: FnMut(&Operation<'a>, &mut Exec, Exec::VirtualAddress),

Runs analyzer to end with any user interaction, producing Cfg.

Maybe to be removed later? User code should ideally not use end_analysis and expect that any pending branches get processed by finish().