scarf::analysis

Trait Analyzer

pub trait Analyzer<'e>: Sized {
    type Exec: ExecutionState<'e>;
    type State: AnalysisState;

    // Provided methods
    fn operation(
        &mut self,
        _control: &mut Control<'e, '_, '_, Self>,
        _op: &Operation<'e>,
    ) { ... }
    fn branch_start(&mut self, _control: &mut Control<'e, '_, '_, Self>) { ... }
    fn branch_end(&mut self, _control: &mut Control<'e, '_, '_, Self>) { ... }
}

The trait that will be implemented by user-side code to hook into FuncAnalysis to retrieve whichever results are wanted.

Most commonly the Analyzer will implement at least the operation() hook, which will be called for each sub-step of each instruction FuncAnalysis is walking through.

During the hooks the Analyzer has access to Control, which is used to read and modify any scarf-side analysis state available to user code, as well as a &mut self reference to the Analyzer structure, which can contain any state the user code needs to be shared across entire analysis. If the user code needs branch-specific state with custom state merging behaviour, Analyzer::State can be used to easily implement the state.

Required Associated Types

type Exec: ExecutionState<'e>

The ExecutionState, CPU architecture that will be used by this analyzer.

type State: AnalysisState

Additional user-defined state that will be merged by analysis when two branches of execution join.

If not needed, DefaultState can be used to get a state which carries no information.

When not using DefaultState, call FuncAnalysis::custom_state to create the FuncAnalysis struct. The state can be read and modified during all Analyzer hooks with Control::user_state.

Provided Methods

fn operation( &mut self, _control: &mut Control<'e, '_, '_, Self>, _op: &Operation<'e>, )

Called for each Operation of disassembled instruction.

This is the main function for user code’s analysis inspection and manipulation.

fn branch_start(&mut self, _control: &mut Control<'e, '_, '_, Self>)

Called each time the analysis starts executing instructions from a new branch.

fn branch_end(&mut self, _control: &mut Control<'e, '_, '_, Self>)

Called each time the analysis has finished a branch, either due to Operation::Jump or Operation::Return.

Not called if user has requested the branch to be stopped through Control::end_branch.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.