scarf
Module cfg

Structs

Cfg

CfgNode

CfgNodeIter

CfgOutEdgesNodes

NodeBorrow

NodeIndex

NodeLink

The address is considered a stable link to a node, while index is invalid if the parent graph’s node_indices_dirty is true.

NodePredecessors

OutEdgeCondition

Predecessors

Enums

CfgOutEdges

Traits

CfgState