scarf::cfg

Struct Cfg

pub struct Cfg<'e, State: CfgState> { /* private fields */ }

Implementations

impl<'e, S: CfgState> Cfg<'e, S>

pub fn new() -> Cfg<'e, S>

pub fn clear(&mut self)

pub fn get(&self, address: Rva) -> Option<&CfgNode<'e, S>>

pub fn get_link(&self, address: Rva) -> Option<NodeLink>

pub fn get_state(&mut self, address: Rva) -> Option<&mut S>

Gets mutable access to state - which is fine as none of the CFG functionality uses it for anything; itโs purely user data.

Shared access can use cfg.get().map(|x| &x.state)

pub fn entry_link(&self) -> &NodeLink

pub fn entry(&self) -> &CfgNode<'e, S>

pub fn get_by_link(&self, link: &NodeLink) -> &CfgNode<'e, S>

pub fn add_node(&mut self, address: Rva, node: CfgNode<'e, S>)

pub fn nodes<'a>(&'a self) -> CfgNodeIter<'a, 'e, S>

pub fn merge_overlapping_blocks(&mut self)

Tries to replace a block with a jump to another one which starts in middle of current.

pub fn calculate_distances(&mut self)

Can only be called on a graph with all node links being addresses of one of the nodes in graph.

pub fn calculate_node_indices(&mut self)

Can only be called on a graph with all node links being addresses of one of the nodes in graph.

pub fn cycles(&mut self) -> Vec<Vec<NodeLink>>

Can only be called on a graph with all node links being addresses of one of the nodes in graph.

pub fn predecessors(&mut self) -> Predecessors

Retreives a structure which can be used to lookup what nodes branch to current node.

pub fn immediate_postdominator<'a>( &'a self, node: NodeIndex<'a, 'e, S>, ) -> Option<NodeBorrow<'a, 'e, S>>

pub fn resolve_cond_jump_operands<F>(&mut self, convert: F)
where F: FnMut(Operand<'e>, Rva, Rva) -> Operand<'e>,

Converts conditions to resolved form (Relative to branch start).

The actual conversion function fn x(condition: Operand, start: Rva, end: Rva) has to be provided by caller.