Enum scarf::Operation

pub enum Operation<'e> {
    Move(DestOperand<'e>, Operand<'e>),
    Call(Operand<'e>),
    Jump {
        condition: Operand<'e>,
        to: Operand<'e>,
    },
    Return(u32),
    Special(SpecialBytes),
    SetFlags(FlagUpdate<'e>),
    ConditionalMove(DestOperand<'e>, Operand<'e>, Operand<'e>),
    Freeze,
    Unfreeze,
    Error(Error),
}

A sub-operation of simulated instruction.

As the analysis walks through code, it translates instructions to zero or more Operations, which are used as input for analysis to decide which branches are to be analyzed, and to update ExecutionState.

User-side code gets gets to intercept any Operation before it is otherwise processed in Analyzer::operation callback, which is the usually where most of the logic for user’s code is.

In Analyzer::operation, skip_operation can be used to prevent the default Operation from passsing through(*). Whereas update, as well as move_resolved, and move_unresolved can be used to add additional Operations to be processed by scarf.

(*) skip_operation and update do not do anything for control flow operations Operation::Jump and Operation::Return. This probably should be fixed, but end_branch and add_branch_with_current_state can be used to work around this limitation.

As Operation is representing CPU instructions without any external state, all Operands and DestOperands in Operation are always unresolved.

Variants

Move(DestOperand<'e>, Operand<'e>)

Set DestOperand to Operand.

Call(Operand<'e>)

Calls the function at Operand.

ExecutionState just implements this as a “Clear all non-preserved registers” operation, but the user Analyzer code can often have reasons to handle calls.

Note that for performance and usability reasons, ExecutionState only clears registers on functions calls. Any memory is assumed to stay unchanged, which sometimes will cause analysis to be confused when a function was called to write something to a pointer that will be read later. For now, these cases have to be handled by the user code if they pop up.

Jump

Jumps to to if condition is nonzero.

If the analysis is able to determine condition to be constant, it will only analyze the branch that is taken. Otherwise both branches will be queued to be analyzed.

Note that the analysis calls resolve_apply_constraints to resolve the condition, which sometimes gives better results, especially with jump conditions, than the regular resolve would.

Fields

condition: Operand<'e>

to: Operand<'e>

Return(u32)

Returns to caller.

The u32 parameter is additional stack bytes popped after the return address has been popped. (E.g. x86 stdcall with 3 arguments would pop 12 bytes)

Analysis will consider this operation a branch end.

Special(SpecialBytes)

Special instructions that are not representable by other Operations. For example, rep mov is represented with Special. The bytes ideally will represent the instruction bytes, but in the end their interpretation is very much up to ExecutionState. rep mov for example, is completely ignored, but the user Analyzer code can recognize the instruction bytes and do its own handling if deemed necessary.

SetFlags(FlagUpdate<'e>)

Set flags based on operation type. While Move(..) could handle this (And it does for odd cases like inc), it would mean generating 5 additional operations for each instruction, so special-case flags.

ConditionalMove(DestOperand<'e>, Operand<'e>, Operand<'e>)

Set DestOperand to Operand if the second Operand evaluates to nonzero.

The conditional moves are generated quite rarely, and in practice they usually cause move of Undefined to the DestOperand.

Freeze

Makes the following Operations until Unfreeze be buffered, resolving input Operands, without mutating the state.

Allows implementing operations that write to several outputs, which are simultaneously used as inputs. For example, long mul, div (Result + modulo), swap, carry add/sub.

Cloning state which is in middle of freeze will not clone any buffered operations; the new copy can be updated but any buffered writes are lost.

Unfreeze

Commits any buffered operations since Freeze was used to buffer them.

Error(Error)

Error - Should assume that no more operations can be decoded from current position.