scarf

Module exec_state

Traits for abstracting over different CPU architecture, and code that can be shared between them.

Main points of interest are ExecutionState for the main architecture trait, and VirtualAddress for the “Integer representing memory address of word size” trait.

Structs

Constraint

The constraint is assumed to be something that can be substituted with 1 if met (so constraint == constval(1)).

Memory

Contains memory state as addr -> value hashmap. The ExecutionState is expected to take care of cases where memory is written with one address and part of it is read at offset address, in practice by splitting accesses to be word-sized, and any misaligned accesses become bitwise and-or combinations.

MemoryValue

As the memory is stored in larger blocks than bytes (In effect ExecutionState words), write to nearby byte will make entire word merge to Undefined. To alleviate this inaccuracy a bit, keep track of bytes that have never been written, and consider value to not apply to these bytes. This also makes smaller-than-word writes bit more efficient when they don’t have to be merged with Mem64[addr] for the untouched bytes.

MergeStateCache

A cache which allows skipping some repeated work during merges.

Statics

X86_64_ARCH_DEFINITION

Traits

Disassembler

Trait for disassembling instructions

ExecutionState

A trait that does (most of) arch-specific state handling.

OperandCtxExtX86

OperandExtX86

VirtualAddress

Either scarf::VirtualAddress in 32-bit or scarf::VirtualAddress64 in 64-bit

Functions

assume_jump_flag

Updates the two states for conditional jump to have state based on whether jump is taken or not.

is_flag_const_constraint

For constraint X, return Y: Assumes that flags are 1bit, which isn’t super set in stone elsewhere (yet) flag == 0 => (flag, 0) flag != 0 => (flag, 1) flag == 1 => (flag, 1)

merge_constraint

resolve_address

Resolves to (base, offset) though base itself may contain offset as well, just tries to avoid interning new constant additions

value_limits

Helper for ExecutionState::value_limits implementations with constraints