scarf::exec_state

Struct Memory

pub struct Memory<'e> { /* private fields */ }

Contains memory state as addr -> value hashmap. The ExecutionState is expected to take care of cases where memory is written with one address and part of it is read at offset address, in practice by splitting accesses to be word-sized, and any misaligned accesses become bitwise and-or combinations.

Implementations

impl<'e> Memory<'e>

pub fn new() -> Memory<'e>

pub fn get(&mut self, base: Operand<'e>, offset: u64) -> Option<MemoryValue<'e>>

Low-level api to read memory that is not aware of how many bytes one Operand represents.

pub fn set(&mut self, base: Operand<'e>, offset: u64, value: MemoryValue<'e>)

Low-level api to write memory that is not aware of how many bytes one Operand represents.

pub fn read( &mut self, ctx: OperandCtx<'e>, base: Operand<'e>, byte_offset: u64, size: MemAccessSize, word_size: MemAccessSize, make_unwritten: impl FnMut(Operand<'e>, u64) -> Operand<'e>, ) -> Option<Operand<'e>>

High-level memory read.

pub fn write( &mut self, ctx: OperandCtx<'e>, value: Operand<'e>, base: Operand<'e>, byte_offset: u64, size: MemAccessSize, word_size: MemAccessSize, )

pub fn is_same(&self, other: &Memory<'e>) -> bool

Cheap check for if self and other share their hashmap.

If true is returned, maps are equal; false can still mean that the maps are equal, it just canโt be cheaply confirmed.

pub fn maybe_convert_immutable(&mut self, limit: usize)

pub fn has_merge_changed(&self, b: &Memory<'e>) -> bool