scarf::operand

Struct Operand

pub struct Operand<'e>(/* private fields */);

Operand is the type of values in scarf.

It is an immutable reference type, free to copy and expected to be passed by value.

Different types of Operands are listed in OperandType enum. The main types of interest are:

Single variables, such as OperandType::Arch, OperandType::Custom.

Constant integers, OperandType::Constant.

Expressions, OperandType::Arithmetic, using Operand as inputs for the expression, and as such, able to have arbitrarily deep tree of subexpressions making up the Operand.

Memory, OperandType::Memory, which is able to have any Operand representing the address.

All Operands are created through OperandContext arena, which interns the created Operands, allowing implementing equality comparison as a single reference equality check. The lifetime 'e refers to this backing memory in OperandContext which the Operand reference points to.

Simplification

Every created Operand which contains an expression is immediately simplified to a form that scarf considers to be the simplest equivalent form of the expression.

For example, subtracting any Operand from itself is simplified to zero, while adding any Operand to itself gets simplfied to multiplication by two:

let ctx = &scarf::OperandContext::new();
// Create an `Operand` representing `rcx` register.
// x86-64 registers are ordered as 'rax rcx rdx rbx  rsp rbp rsi rdi  r8 r9 r10...
// As such, index of rcx is 1.
let rcx = ctx.register(1);
// Create an `Operand` representing `rcx - rcx` expression.
let sub_result = ctx.sub(rcx, rcx);
// Create an `Operand` representing constant 0.
let zero = ctx.constant(0);
// Subtraction result is also zero
assert_eq!(sub_result, zero);

// And `rcx + rcx` is same as `rcx * 2`
let add_result = ctx.add(rcx, rcx);
// Most arithmetic `Operand` creation functions have a variation
// which should be preferred when one of the operands is known to
// be constant.
// This is more concise and better for performance than the equivalent
// `ctx.mul(rcx, ctx.constant(2))`.
let mul_result = ctx.mul_const(rcx, 2);
assert_eq!(add_result, mul_result);

// We can also have an `Operand` expression for equivalency:
// `ArithOpType::Equal` evaluates to either 0 or 1 for the result.
let eq_result = ctx.eq(add_result, mul_result);
// Equivalent to `ctx.constant(1)`, but effectively free to evaluate.
let one = ctx.const_1();
assert_eq!(eq_result, one);

All Operands, and as such, arithmetic expressions of the Operands are considered 64-bit values. In general, arithmetic expressions wrap on overflow, and arithmetic smaller than 64-bit is represented by doing bitwise AND to only keep the low bits.

Note: Overflowing shifts (Shifting by more than 64 bits) are defined to result in zero, instead of the behaviour of x86 shift and rotate instructions on general-purpose registers, where shift count is taken as modulo 64. Scarf represents those instructions as rax << (count & 63), and users generally do not have to think about this.

If scarf is given excessively complex arithmetic to simplify – which is often seen in highly unrolled hash functions, scarf rather gives up than never finishes the simplification.

Canonicalization guarantees

Simplification tries to convert any equivalent expression to a single ‘canonical’ form, to make equality comparisons useful. As the simplification rules get improved, this may result in user code which matches against a specific form of arithmetic expression to break.

This does mean that updating scarf dependency can bring surprising breaking changes! Try to have tests on any code that has to match on Operands.

The following canonicalizations are guaranteed to be stable across scarf changes, to allow code to be simpler and faster by not having to consider expression forms that are never created.

Any chain of an operation, where the operation is commutative (Swapping left/right does not affect result; Mul/And/Or/Xor) will be simplified to have single operand on right and rest of the chain on left:

(((a ^ b) ^ c) ^ d) ^ e.

Expressions chains of multiple combined Add/Sub are simplified similar to above rule:

(((a + b) - c) + d) + e

No guarantee how adds and subs are ordered within the chain.

If any of the above commutative/add-sub chains has a constant, it is guaranteed to be the outermost right operand:

(a & b) & ffff

Equality expressions containing a non-zero constant always have the constant alone on right, with rest of the expression shuffled to left. If the constant is zero, and there is no subtraction on left side, the zero is also guaranteed to be on left.

That is, x - y == 0 is simplified to either x == y or y == x instead, while 0 == x is guaranteed be simplified to x == 0.

Consider using Operand::either for cases where the result you want may be on either left- or right-hand operand of the expression.

Matching

The most comprehensive way to match against different values of Operand is to match on the OperandType enum, returned from Operand::ty.

However, most of the time this leads to quite verbose code, especially when you want to try match against just a single variant of arithmetic expression or constant.

Operand includes a large amount of matching methods, starting with if_, such as if_constant, returning Option<u64>, if_arithmetic_add, returning Option<(Operand<'e>, Operand<'e>)>, if_memory, returning Option<&'e MemAccess<'e>>.

Using the matching methods is usually a lot more concise than using match on OperandType.

Note that the memory matching methods effectively have three tiers:

if_memory, returning MemAccess if the Operand is memory operand at all.

if_mem64, if_mem32, and others, which still return MemAccess, but check that mem.size is of expected size

if_mem64_offset, if_mem32_offset, and others, which check both MemAccess size. and specific constant offset, commonly useful when matching on a structure field.

They return Operand for the base memory address.

Implementations

impl<'e> Operand<'e>

pub fn ty(self) -> &'e OperandType<'e>

Returns reference to actual enum type carrying data of this Operand.

pub fn relevant_bits(self) -> Range<u8>

Returns what bits in this value are not guaranteed to be zero.

End cannot be larger than 64.

Can be also seen as trailing_zeros .. 64 - leading_zeros range

pub fn relevant_bits_mask(self) -> u64

pub fn if_constant(self) -> Option<u64>

Returns Some(c) if self.ty is OperandType::Constant(c)

pub fn if_custom(self) -> Option<u32>

Returns Some(c) if self.ty is OperandType::Custom(c)

pub fn if_register(self) -> Option<u8>

Returns Some(r) if self.ty is OperandType::Arch(r) with r.value() being less than 256.

pub fn if_arch_id(self) -> Option<u32>

pub fn x86_if_flag(self) -> Option<Flag>

Returns Some(f) if self.ty is OperandType::Flag(f)

pub fn if_memory(self) -> Option<&'e MemAccess<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem)

pub fn if_mem64(self) -> Option<&'e MemAccess<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem64

pub fn if_mem32(self) -> Option<&'e MemAccess<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem32

pub fn if_mem16(self) -> Option<&'e MemAccess<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem16

pub fn if_mem8(self) -> Option<&'e MemAccess<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem8

pub fn if_mem64_offset(self, offset: u64) -> Option<Operand<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem64

pub fn if_mem32_offset(self, offset: u64) -> Option<Operand<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem32

pub fn if_mem16_offset(self, offset: u64) -> Option<Operand<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem16

pub fn if_mem8_offset(self, offset: u64) -> Option<Operand<'e>>

Returns Some(mem) if self.ty is OperandType::Memory(ref mem) and mem.size == MemAccessSize::Mem8

pub fn if_arithmetic( self, ty: ArithOpType, ) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic { ty == ty }

pub fn if_arithmetic_any(self) -> Option<&'e ArithOperand<'e>>

Returns Some(arith) if self.ty is OperandType::Arithmetic(arith).

pub fn is_arithmetic(self, ty: ArithOpType) -> bool

Returns true if self.ty is OperandType::Arithmetic { ty == ty }

pub fn if_arithmetic_add(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Add(left, right))

pub fn if_arithmetic_sub(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Sub(left, right))

pub fn if_arithmetic_mul(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Mul(left, right))

pub fn if_arithmetic_mul_high(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::MulHigh(left, right))

pub fn if_arithmetic_eq(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Equal(left, right))

pub fn if_arithmetic_gt(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::GreaterThan(left, right))

pub fn if_arithmetic_and(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::And(left, right))

pub fn if_arithmetic_or(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Or(left, right))

pub fn if_arithmetic_xor(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Xor(left, right))

pub fn if_arithmetic_lsh(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Lsh(left, right))

pub fn if_arithmetic_rsh(self) -> Option<(Operand<'e>, Operand<'e>)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Rsh(left, right))

pub fn if_arithmetic_with_const( self, arith: ArithOpType, ) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic({ ty: arith, left, right: OperandType::Constant(right) })

pub fn if_add_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Add(left, OperandType::Constant(right)))

pub fn if_sub_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Sub(left, OperandType::Constant(right)))

pub fn if_mul_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Mul(left, OperandType::Constant(right)))

pub fn if_mul_high_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::MulHigh(left, OperandType::Constant(right)))

pub fn if_eq_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Equal(left, OperandType::Constant(right)))

pub fn if_gt_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::GreaterThan(left, OperandType::Constant(right)))

pub fn if_and_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::And(left, OperandType::Constant(right)))

pub fn if_or_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Or(left, OperandType::Constant(right)))

pub fn if_xor_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Xor(left, OperandType::Constant(right)))

pub fn if_lsh_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Lsh(left, OperandType::Constant(right)))

pub fn if_rsh_with_const(self) -> Option<(Operand<'e>, u64)>

Returns Some((left, right)) if self.ty is OperandType::Arithmetic(ArithOpType::Rsh(left, OperandType::Constant(right)))

pub fn if_sign_extend( self, ) -> Option<(Operand<'e>, MemAccessSize, MemAccessSize)>

Returns Some((val, from, to)) if self.ty is OperandType::SignExtend(val, from, to)

pub fn if_select(self) -> Option<(Operand<'e>, Operand<'e>, Operand<'e>)>

Returns Some((condition, val_true, val_false)) if self.ty is OperandType::Select(condition, val_true, val_false)

pub fn if_and_masked_register(self) -> Option<(u8, u64)>

Returns Some((register, constant)) if operand is an and mask of register with constant.

Useful for detecting 32-bit register which is represented as Register(r) & ffff_ffff.

pub fn is_undefined(self) -> bool

Returns true if self.ty() == OperandType::Undefined.

pub fn contains_undefined(self) -> bool

Returns true if self or any child operand is Undefined.

pub fn contains_memory(self) -> bool

Returns true if self or any child operand is Memory.

pub fn and_masked(this: Operand<'e>) -> (Operand<'e>, u64)

Returns (other, constant) if operand is an and mask with constant, or just (self, u64::MAX) otherwise.

pub fn either<F, T>( a: Operand<'e>, b: Operand<'e>, f: F, ) -> Option<(T, Operand<'e>)>
where F: FnMut(Operand<'e>) -> Option<T>,

If either of a or b matches the filter-map f, return the mapped result and the other operand.

pub fn iter(self) -> Iter<'e>

Returns an type implementing Iterator, yielding self, followed by other Operands making up self in some unspecified order. If the same Operand is included in multiple branches of expression, it will be yielded mutiple times.

Note that this needs to do a memory allocation to keep track of the iteration state, it is often better to match against expected form of expression than use iter to try find a specific suboperand.

Additionally, scarf allows constructing Operand with N^2 (not unique) suboperands in just N operations. This is often seen in hash functions, which shuffle bits of individual inputs without exactly discarding any of them. While scarf handles these massive Operands fine, trying to iterate through them (Or even debug print them) will take excessive amounts of time. Code using iter that may be used in hash functions needs to have some limit after which iteration will be stopped.

pub fn iter_no_mem_addr(self) -> IterNoMemAddr<'e>

Returns an Iterator type similar to iter, expect that it does not yield operands or their suboperands making up OperandType::Memory addresses. The operand with type OperandType::Memory is still yielded, it just is not recursed into.

Same caveats that iter has apply here too.

pub fn to_self_owned(self) -> SelfOwnedOperand

Converts this reference to &'e OperandContext to a struct which carries the backing OperandContext with it.

As this means that a new OperandContext will be allocated and all operands this refers to will be copied there, the operation can be relatively slow.

pub fn hash_by_address(self) -> OperandHashByAddress<'e>

Returns this operand wrapped in a newtype that implements Hash by hashing the address of this reference.

pub fn const_offset( oper: Operand<'e>, ctx: OperandCtx<'e>, ) -> Option<(Operand<'e>, u64)>

👎Deprecated

Matches against:

(x + c), returning (x, c), (x - c), returning (x, -c), c, returning (0, c)

Probably should be removed.