scarf::operand

Struct OperandContext

pub struct OperandContext<'e> { /* private fields */ }

An arena for allocating and interning Operands.

Usually the type alias to reference OperandCtx is used instead. A variable of type OperandContext / OperandCtx is usually named ctx.

OperandContext is used to create Operands, returning reference to already existing Operand if it was already created, allowing equality comparisons be done with a simple pointer equality check.

Additionally OperandContext makes sure that all created Operands are simplified and canonicalized. For example, rcx - rcx is simplified to a constant 0, while rax + rcx and rcx + rax return references to the same Operand. While the choice between rax + rcx or rcx + rax for canonical form is not specified, some canonicalizations are guaranteed, and user code may rely on them to simplify matching of Operands.

Most of the functions of OperandContext are for creating an Operand with specific OperandType, though there is also a mem_access to create MemAccess (As it also has canonicalization requirements), and few general-use functions

OperandContext lifetime 'e

The 'e lifetime passed around everywhere in scarf ultimately refers to Operands, which are a reference to &'e OperandContext<'e>, usually shortened with a type alias to [OperandCtx<'e>].

'e is an invariant lifetime, which generally makes it impossible to mix allocations from two different OperandContexts. However, if you have OperandCtx<'static> or create two separate OperandContexts in same scope, Rust allows assigning the same lifetime to them, allowing the allocations to be mixed. While this is not expected to cause memory safety issues, it will make scarf produce nonsensical results as equality comparisons are no longer reliable. Try to avoid cases where there are two OperandContext with same lifetime.

A common use case for multiple OperandContexts is to use a shorter-lived one to analyze functions (Which can allocate tens or hundrerds of thousands Operands total), that will be throwns away to release memory afterwards, and a longer-lived OperandContext to hold Operand results that you are interested in. Operands can be copied across OperandContext with copy_operand.

Cheap-to-access Operands

In general, creating an Operand requires at least a hash table lookup / insertion to make sure Operands are equal if their references are equal. As such, avoiding doing repeated hash table lookups by storing the returned Operand is preferrable. However, few of the most common Operands are already cached by OperandContext, and any code having OperandContext reference can rely on retreiving these Operands being a single memory access.

The Operands which are guaranteed to be cached are:

OperandType::Arch(n), for n.value() < 16

OperandType::Constant(n), for 0, 1, 2, 4, 8, through ctx.const_0() and other functions.

Some code uses this to micro-optimize cases where you want to check if Operand equals to constant zero/one or constant register. op.if_constant() == 0 has to do two comparisons and two memory reads: One to verify OperandType::Constant variant, second to verify that the constant is zero. op == ctx.const_0() does just a single memory read to get the cached const_0, and a single pointer comparison.

Implementations

impl<'e> OperandContext<'e>

pub fn new() -> OperandContext<'e>

Creates a new OperandContext.

This will immediately spend some time interning the most common Operands that will be cheap to access later, maybe making this relatively slow operation compared to rest of scarf. See docs on guarantees for Operands that are guaranteed to be cheap to access afterwards.

pub fn new_with_arch(arch_def: &'e ArchDefinition<'e>) -> OperandContext<'e>

pub fn copy_operand<'other>(&'e self, op: Operand<'other>) -> Operand<'e>

Copies an operand referring to some other OperandContext to this OperandContext and returns the copied reference.

pub fn const_0(&'e self) -> Operand<'e>

Cheap access to a small constant.

pub fn const_1(&'e self) -> Operand<'e>

Cheap access to a small constant.

pub fn const_2(&'e self) -> Operand<'e>

Cheap access to a small constant.

pub fn const_4(&'e self) -> Operand<'e>

Cheap access to a small constant.

pub fn const_8(&'e self) -> Operand<'e>

Cheap access to a small constant.

pub fn flag_z(&'e self) -> Operand<'e>

Cheap access to Flag::Zero operand.

pub fn flag_c(&'e self) -> Operand<'e>

Cheap access to Flag::Carry operand.

pub fn flag_o(&'e self) -> Operand<'e>

Cheap access to Flag::Overflow operand.

pub fn flag_s(&'e self) -> Operand<'e>

Cheap access to Flag::Sign operand.

pub fn flag_p(&'e self) -> Operand<'e>

Cheap access to Flag::Parity operand.

pub fn flag_d(&'e self) -> Operand<'e>

Cheap access to Flag::Direction operand.

pub fn flag(&'e self, flag: Flag) -> Operand<'e>

Cheap access to any OperandType::Flag operand.

pub fn register(&'e self, num: u8) -> Operand<'e>

Returns OperandType::Arch(num) operand.

If num is less than 16, the access is guaranteed to be cheap.

pub fn arch(&'e self, num: u32) -> Operand<'e>

pub fn custom(&'e self, value: u32) -> Operand<'e>

Returns OperandType::Custom(value) operand.

Custom operands are guaranteed to be never generated by scarf, allowing user code use them for arbitrary unknown variables of any meaning.

pub fn constant(&'e self, value: u64) -> Operand<'e>

Returns OperandType::Constant(value) operand.

pub fn arithmetic( &'e self, ty: ArithOpType, left: Operand<'e>, right: Operand<'e>, ) -> Operand<'e>

Returns Operand for any arithmetic operation.

pub fn arithmetic_masked( &'e self, ty: ArithOpType, left: Operand<'e>, right: Operand<'e>, mask: u64, ) -> Operand<'e>

Returns Operand for any arithmetic operation, which will be then masked with a bitwise AND of mask.

This may have better performance than the equivalent ctx.and_const(ctx.arithmetic(ty, left, right), mask), as it is able to avoid allocating expressions that will be then immediately discarded.

pub fn float_arithmetic( &'e self, ty: ArithOpType, left: Operand<'e>, right: Operand<'e>, size: MemAccessSize, ) -> Operand<'e>

Returns Operand for a float arithmetic of any operation.

pub fn add(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left + right.

pub fn sub(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left - right.

pub fn mul(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left * right.

pub fn mul_high(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for high 64 bits of 128-bit result of left * right.

pub fn signed_mul( &'e self, left: Operand<'e>, right: Operand<'e>, _size: MemAccessSize, ) -> Operand<'e>

Returns Operand for signed left * right.

pub fn div(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left / right.

pub fn modulo(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left % right.

pub fn and(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left & right.

pub fn or(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left | right.

pub fn xor(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left ^ right.

pub fn lsh(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left << right.

pub fn rsh(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left >> right.

pub fn arithmetic_right_shift( &'e self, left: Operand<'e>, right: Operand<'e>, size: MemAccessSize, ) -> Operand<'e>

pub fn eq(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left == right.

pub fn neq(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for left != right.

pub fn gt(&'e self, left: Operand<'e>, right: Operand<'e>) -> Operand<'e>

Returns Operand for unsigned left > right.

Less than can be implemented by swapping left and right.

Greater than or equal can be implemented by using ctx.or(ctx.gt(a, b), ctx.eq(a, b))

pub fn gt_signed( &'e self, left: Operand<'e>, right: Operand<'e>, size: MemAccessSize, ) -> Operand<'e>

Returns Operand for signed left > right.

pub fn add_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left + right.

pub fn sub_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left - right.

pub fn sub_const_left(&'e self, left: u64, right: Operand<'e>) -> Operand<'e>

Returns Operand for left - right.

pub fn mul_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left * right.

pub fn and_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left & right.

pub fn or_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left | right.

pub fn xor_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left ^ right.

pub fn lsh_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left << right.

pub fn lsh_const_left(&'e self, left: u64, right: Operand<'e>) -> Operand<'e>

Returns Operand for left << right.

pub fn rsh_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left >> right.

pub fn eq_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left == right.

pub fn neq_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for left != right.

Currently this is represented as (left == right) == 0.

pub fn gt_const(&'e self, left: Operand<'e>, right: u64) -> Operand<'e>

Returns Operand for unsigned left > right.

pub fn gt_const_left(&'e self, left: u64, right: Operand<'e>) -> Operand<'e>

Returns Operand for unsigned left > right.

pub fn mem64(&'e self, base: Operand<'e>, offset: u64) -> Operand<'e>

Returns Operand for MemAccessSize::Mem64 memory read, using given base and offset.

pub fn mem32(&'e self, base: Operand<'e>, offset: u64) -> Operand<'e>

Returns Operand for MemAccessSize::Mem32 memory read, using given base and offset.

pub fn mem16(&'e self, base: Operand<'e>, offset: u64) -> Operand<'e>

Returns Operand for MemAccessSize::Mem16 memory read, using given base and offset.

pub fn mem8(&'e self, base: Operand<'e>, offset: u64) -> Operand<'e>

Returns Operand for MemAccessSize::Mem8 memory read, using given base and offset.

pub fn mem64c(&'e self, offset: u64) -> Operand<'e>

Returns Operand for Mem64 with constant address.

pub fn mem32c(&'e self, offset: u64) -> Operand<'e>

Returns Operand for Mem32 with constant address.

pub fn mem16c(&'e self, offset: u64) -> Operand<'e>

Returns Operand for Mem16 with constant address.

pub fn mem8c(&'e self, offset: u64) -> Operand<'e>

Returns Operand for Mem8 with constant address.

pub fn mem_any( &'e self, size: MemAccessSize, addr: Operand<'e>, offset: u64, ) -> Operand<'e>

Creates Operand referring to memory from parts that make up MemAccess.

pub fn memory(&'e self, mem: &MemAccess<'e>) -> Operand<'e>

Creates Operand referring to memory from MemAccess.

pub fn mem_access( &'e self, base: Operand<'e>, offset: u64, size: MemAccessSize, ) -> MemAccess<'e>

Creates MemAccess from base, offset, and size.

pub fn mem_access8(&'e self, base: Operand<'e>, offset: u64) -> MemAccess<'e>

Shortcut for ctx.mem_access(base, offset, MemAccessSize::Mem8)

pub fn mem_access16(&'e self, base: Operand<'e>, offset: u64) -> MemAccess<'e>

Shortcut for ctx.mem_access(base, offset, MemAccessSize::Mem16)

pub fn mem_access32(&'e self, base: Operand<'e>, offset: u64) -> MemAccess<'e>

Shortcut for ctx.mem_access(base, offset, MemAccessSize::Mem16)

pub fn mem_access64(&'e self, base: Operand<'e>, offset: u64) -> MemAccess<'e>

Shortcut for ctx.mem_access(base, offset, MemAccessSize::Mem16)

pub fn mem_add_op( &'e self, mem: &MemAccess<'e>, value: Operand<'e>, ) -> Operand<'e>

Creates Operand representing mem.address_op() + value without creating intermediate mem.address_op()

pub fn mem_sub_op( &'e self, mem: &MemAccess<'e>, value: Operand<'e>, ) -> Operand<'e>

Creates Operand representing mem.address_op() - value without creating intermediate mem.address_op()

pub fn mem_add_const_op( &'e self, mem: &MemAccess<'e>, value: u64, ) -> Operand<'e>

Creates Operand representing mem.address_op() + value without creating intermediate mem.address_op()

pub fn mem_sub_const_op( &'e self, mem: &MemAccess<'e>, value: u64, ) -> Operand<'e>

Creates Operand representing mem.address_op() - value without creating intermediate mem.address_op()

pub fn sign_extend( &'e self, val: Operand<'e>, from: MemAccessSize, to: MemAccessSize, ) -> Operand<'e>

Creates an OperandType::SignExtend Operand.

pub fn select( &'e self, condition: Operand<'e>, val_true: Operand<'e>, val_false: Operand<'e>, ) -> Operand<'e>

Creates an OperandType::Select Operand. `if condition != 0 { val_true } else { val_false }

pub fn new_undef(&'e self) -> Operand<'e>

Creates a new OperandType::Undefined(id) Operand, where the id is unique until all currently ongoing FuncAnalysis::analyze calls have returned.

See FuncAnalysis docs for more information about how OperandType::Undefined is used by analysis to determine which branches have to be rechecked.

pub fn transform<F>( &'e self, oper: Operand<'e>, depth_limit: usize, f: F, ) -> Operand<'e>
where F: FnMut(Operand<'e>) -> Option<Operand<'e>>,

Walks through sub-operands of an Operand, calling the provided callback that can choose to replace any of them with a new Operand by returning Some().

pub fn substitute( &'e self, oper: Operand<'e>, val: Operand<'e>, with: Operand<'e>, depth_limit: usize, ) -> Operand<'e>

pub fn interned_counts(&self) -> InternedCounts

Gets counts of different types of operands interned.

Intented for debug / diagnostic info.

pub fn truncate(&'e self, operand: Operand<'e>, size: u8) -> Operand<'e>

Returns operand limited to low size bits.

pub fn normalize_32bit(&'e self, value: Operand<'e>) -> Operand<'e>

Converts operand to โ32-bit normalized formโ

This function exists for 32-bit ExecutionState, which wants to mainly deal with Operands without caring about their high 32 bits.

This could be achieved with just ctx.and_const(value, 0xffff_ffff) on every Operand, but these masked values would then need to be constantly unwrapped. For values that donโt use the entire 32 bits, such as Mem32[400] << 8, the mask would become just 0xffff_ff00 after simplification, and determining if the mask is just used intentionally or if it is just to mask out high bits becomes more complicated.

So instead of always zeroing high bits, the following kinds of Operands have their โ32-bit normalized formโ be as is, without an bitwise AND mask, even if rest of scarf can consider their high bits to possibly have a meaning. Operands for which scarf can prove low 32 bits be always zero will become zero though, even if they would otherwise be included in the rules below. (E.g. Mem32[x] << 0x20)

Any โpure 64-bit variableโ; Arch, Undefined, Custom

Additions, subtractions, consisting only 32-bit normalized Operands

Multiplications, left shifts consisting 32-bit normalized Operand and a constant

Most XORs that have both halves 32-bit normalized. (Details left loosely defined for now though)

OperandType::ArithmeticFloat

If the Operand is AND 0xffff_ffff masked, but the value being masked meets above conditions, the mask is removed.

If the Operand is AND 0xffff_ffff masked addition or subtraction, with some of the terms meeting the above conditions, those terms are moved outside of the AND mask. E.g. (rax + (rcx ^ rdx)) & ffff_ffff becomes rax + ((rcx ^ rdx) & ffff_ffff).

In other cases bitwise AND mask 0xffff_ffff is applied. (Which does not change the input value if all its high bits are already zero)

In the end, it is intended, though not necessarily guaranteed in complex cases, that ctx.normalize_32bit(x) == ctx.normalize_32bit(x & ffff_ffff).