scarf::operand

Type Alias OperandCtx

pub type OperandCtx<'e> = &'e OperandContext<'e>;

Convenience alias for OperandContext reference that avoids having to type the 'e lifetime twice.

This is the type used in all function parameters, OperandContext is only needed when constructing it.