scarf

Struct BinaryFile

pub struct BinaryFile<Va: VirtualAddress> {
    pub base: Va,
    /* private fields */
}

Contains the binary that is to be analyzed, loaded to memory.

Not much more than a Vec<BinarySection>, and functions to read the correct section in different ways, when given VirtualAddress.

BinaryFile is often passed as by reference &'e BinaryFile, using same lifetime as OperandContext. This would not be required, as BinaryFile does not hold any references to OperandContext, the lifetime was chosen to be same to keep all code interfacing with scarf simpler, only requiring a single lifetime being passed around. This unfortunately does require BinaryFile to be created so that it outlives the OperandContext passed to FuncAnalysis, which usually should not be a too big of a problem.

parse and parse_x86_64 can be used to load a BinaryFile from Windows PE executable, raw_bin can be used to create a BinaryFile from arbitrary sections.

Fields
base: Va
Implementations

impl<Va: VirtualAddress> BinaryFile<Va>

pub fn base(&self) -> Va

pub fn rva_32(&self, address: Va) -> u32

Returns (address - self.base()) as u32.

Panics on overflow when overflow checks are enabled. If the address is known to be in one of the sections, at least on Windows any PE file will have offset that should fit in u32 even if truncated.

pub fn try_rva_32(&self, address: Va) -> Option<u32>

Returns (address - self.base()) as u32, or None if address is less than self.base() or the difference does not fit in u32.

pub fn rva_64(&self, address: Va) -> u64

Returns (address - self.base()) as u64.

Panics on overflow when overflow checks are enabled.

pub fn try_rva_64(&self, address: Va) -> Option<u64>

Returns (address - self.base()) as u64, or None if address is less than self.base().

pub fn code_section(&self) -> &BinarySection<Va>

Returns the section named โ.textโ.

Panics if this section for some reason does not exist.

[code_sections] should be preferred if the user code wants to handle possibility of multiple executable sectionsโฆ But currently its implementation is bad and just returns section named โ.textโ, if it existsโฆ

pub fn sections<'a>( &'a self, ) -> impl Iterator<Item = &'a BinarySection<Va>> + 'a

Returns iterator, which yields all [BinarySection]s of this BinaryFile.

pub fn code_sections<'a>( &'a self, ) -> impl Iterator<Item = &'a BinarySection<Va>> + 'a

pub fn section(&self, name: &[u8; 8]) -> Option<&BinarySection<Va>>

Returns section by name, if it exists.

pub fn section_by_addr(&self, address: Va) -> Option<&BinarySection<Va>>

Returns a section containing the address, if it exists

pub fn slice_from(&self, range: Range<u32>) -> Result<&[u8], OutOfBounds>

Range is relative from base

pub fn slice_from_address( &self, start: Va, len: u32, ) -> Result<&[u8], OutOfBounds>

Receives a slice of data from address with length len, or Err(OutOfBounds) if there is no single section containing the entire slice.

pub fn slice_from_address_to_end(&self, start: Va) -> Result<&[u8], OutOfBounds>

Receives a slice of data from address to end of the respective section.

pub fn read_u8(&self, addr: Va) -> Result<u8, OutOfBounds>

Searches for section containing addr, and reads a little-endian u8 from there.

pub fn read_u16(&self, addr: Va) -> Result<u16, OutOfBounds>

Searches for section containing addr, and reads a little-endian u16 from there.

pub fn read_u32(&self, addr: Va) -> Result<u32, OutOfBounds>

Searches for section containing addr, and reads a little-endian u32 from there.

pub fn read_u64(&self, addr: Va) -> Result<u64, OutOfBounds>

Searches for section containing addr, and reads a little-endian u64 from there.

pub fn read_address(&self, addr: Va) -> Result<Va, OutOfBounds>

Searches for section containing addr, and reads a VirtualAddress (Same size as the addresses of this BinaryFile are) from there.

pub fn set_relocs(&mut self, relocs: Vec<Va>)