scarf

Module analysis

Contains FuncAnalysis and related types and traits.

Additionally miscellaneous heuristic functions to find function entries and relocations of BinaryFile are included.

Structs

CfgState

Control

Provides access to analysis state when in a Analyzer callback.

DefaultState

A no-op zero-sized type implementing AnalysisState that can be used if there is no need for user-side mergeable state.

FuncAnalysis

The analysis runner of scarf.

FuncCallPair

FuncPtrPair

RelocValues

Enums

Error

Errors from disassembly (Operation generation)

Traits

AnalysisState

A trait for additional branch-specific analysis state that will be merged by the analysis when two different branches join.

Analyzer

The trait that will be implemented by user-side code to hook into FuncAnalysis to retrieve whichever results are wanted.

Functions

find_functions

Attempts to find functions of a binary, accepting ones that are called/long jumped to from somewhere. Returns addresses relative to start of code section.

find_functions_with_callers

Sorted by callee, so looking up callers can be done with bsearch

find_functions_with_callers_x86

find_functions_with_callers_x86_64

find_relocs

find_relocs_x86

find_relocs_x86_64

find_switch_tables

relocs_with_values

The returned array is sorted by value.

Type Aliases

Cfg