scarf::exec_state
Function assume_jump_flag

pub fn assume_jump_flag<'e, E>(
    jump_state: &mut E,
    no_jump_state: &mut E,
    condition: Operand<'e>,
    condition_resolved: Operand<'e>,
)
where
    E: ExecutionState<'e>,

Updates the two states for conditional jump to have state based on whether jump is taken or not.

Caller must make sure that condition_resolved == state.resolve(condition) (Or resolve_apply_constraints).

Maybe does bit too much x86-specific treatment?