Module scarf::operand

Operand and its supporting types.

Main types here are OperandContext for the arena all operands are allocated in, and Operand, scarf’s ‘value/variable/expression’ type.

Structs

ArithOperand

InternedCounts

Contains counts of Operands interned by a single OperandContext.

Iter

IterNoMemAddr

MemAccess

Operand

Operand is the type of values in scarf.

OperandContext

An arena for allocating and interning Operands.

OperandHashByAddress

Wrapper around Operand which implements Hash on the interned address. Separate struct since hashing by address gives hashes that aren’t stable across executions or even separate OperandContexts.

SelfOwnedOperand

A single operand which carries all of its data with it.

UndefinedId

Newtype to distinguish each OperandType::Undefined from each other.

Enums

ArithOpType

Enum of arithmetic operations used by scarf Operands. Used by OperandType::Arithmetic and OperandType::ArithmeticFloat, which in turn contain ArithOperand, containing this enum.

Flag

Enumeration of arithmetic result flags of CPU.

MemAccessSize

MemAccessSize is enum for choosing between 8/16/32/64 bit size in scarf.

OperandType

Different values an Operand can hold.

Type Aliases

OperandCtx

Convenience alias for OperandContext reference that avoids having to type the 'e lifetime twice.